% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffMap.R
\name{DiffMap}
\alias{DiffMap}
\title{Compute diffusion coordinates 
 
Computes diffusion coordinates}
\usage{
DiffMap(data, e = LAFON(data), a = 1)
}
\arguments{
\item{data}{raw data for which diffusion coordinates are calculated}

\item{e}{gaussian kernel width. Optimize with either LAFON, SINGER, or custom function}

\item{a}{number of steps in the diffusion Markov chain}
}
\description{
Compute diffusion coordinates 
 
Computes diffusion coordinates
}

