% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffMap.R
\name{LAFON}
\alias{LAFON}
\title{e-value estimation function
 
Computes best e-value according to Lafon criterion}
\usage{
LAFON(data)
}
\arguments{
\item{data}{raw data for which diffusion coordinates are calculated}
}
\description{
e-value estimation function
 
Computes best e-value according to Lafon criterion
}

