% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffMap.R
\name{SINGER}
\alias{SINGER}
\title{e-value estimation function
 
Computes best e-value according to Singer criterion}
\usage{
SINGER(data, start, end)
}
\arguments{
\item{data}{raw data for which diffusion coordinates are calculated}
}
\description{
e-value estimation function
 
Computes best e-value according to Singer criterion
}

